Workaround SPI issue on Pi
After reboot the SPI interface on the Pi seems to be enabled yet
no messages arrive from the card reader. The workaround seems to
be disabling SPI using `raspi-config` and enabling it again before
starting the service.

I tried doing this using a `ExecStartPre=` flag in systemd but
this didn't work.