Add conditional start for radiodan services
These `systemd` services are configured to only start when a [certain path exists][1]:

> With ConditionPathExists= a file existence condition is checked before
> a unit is started. If the specified absolute path name does not exist,
> the condition will fail.

The new default configuration will create all the paths requried to start
everything. Simply rename or delete the files to prevent the services
from running.

The next stage will be to create a process that manages this file
creation, via a web interface.

[1]:https://www.freedesktop.org/software/systemd/man/systemd.unit.html