Add services setup service
A service that manages services!

Building on the [conditional service starting][1], this process
manipulates that conditionality, and puts a UI on top of the system.

Once changes have been saved, your pi will require a reboot.

The next stage will be to manage which apps are running on the system.

[1]:https://github.com/andrewn/neue-radio/pull/26