Configures pulseaudio in system mode
This enables pulseaudio as a systemd service that is started on boot.
This mode is designed for embedded systems that aren't really
multi-user [1]. This is a workaround for a bug where two instances of
pulseaudio are starting for the Pi user causing problems such as not
being able to mute or set the sound level.

[1] https://www.freedesktop.org/wiki/Software/PulseAudio/Documentation/User/SystemWide/