Dynamic serial service
Creates a serial port manager that can open and close several serial
port connections in response to commands over the WebSocket broker. Once
a serial port is opened, data recieved is emitted to subscribers over
the WebSocket and string data can be sent through the port.
node-serialport is used to provide the underlying connection.