Add check for pi platform when adding support apps
A very naive approach. For the purposes of the hackday, we are
assuming that whenever this is being run on Linux, then this is the
Raspberry Pi, and therefore will have access to the setup and debug
services.

This is to prevent new developers not realising why certain apps are not
avaiable when running on their (presumably) MacBook Pro.